+++ CHANGES.md
@@ -1,3 +1,8 @@
+v1.0.0
+
+- removed WIP `new_hconsed` macro (I found no way to make it generic enough)
+- renamed the factory creation macro `new_consign` to just `consign`
+
 v0.9.11
 
 - more homogeneous struct/trait naming